% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{calc_points}
\alias{calc_points}
\title{Help function to calculate points for plot function}
\usage{
calc_points(data, age, number_timesteps, days)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble}}

\item{age}{Range of age}

\item{number_timesteps}{Width of domain}

\item{days}{Number of days over which we will take the average}
}
\value{
Tibble containing points
}
\description{
Help function to calculate points for plot function
}
