% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_death.R
\name{death_count}
\alias{death_count}
\title{Calculate death count by age and gender}
\usage{
death_count(data, age, sex)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble} with RKI data}

\item{age}{age group of which the death count is calculated}

\item{sex}{sex of which the death count is calculated}
}
\value{
death count
}
\description{
Calculate death count by age and gender
}
