% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\name{filter_date}
\alias{filter_date}
\title{Filter tibble with RKI data by date}
\usage{
filter_date(data, from_date, to_date)
}
\arguments{
\item{data}{Tibble with data}

\item{from_date}{From date (inclusive) in YYYYMMDD format}

\item{to_date}{To date (inclusive) in YYYYMMDD format}
}
\value{
a filtered \link[tibble:tibble-package]{tibble}
}
\description{
This function can be used to restrict a tibble with the data by RKI to
the specified time range.
}
\examples{
filter_date(covrki, 20210101, 20210131)
}
