% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_death.R
\name{infection_count}
\alias{infection_count}
\title{Calculate infection count by age and gender}
\usage{
infection_count(data, age, sex)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble} with RKI data}

\item{age}{age group of which the infection count is calculated}

\item{sex}{sex of which the infection count is calculated}
}
\value{
infection count
}
\description{
Calculate infection count by age and gender
}
