% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_rate_death}
\alias{plot_rate_death}
\title{Plot the evolution of the rate of death}
\usage{
plot_rate_death(data, age, number_timesteps, days)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble} with RKI data}

\item{age}{Range of age of subgroup}

\item{number_timesteps}{Time span}

\item{days}{Number of days over which the average is taken}
}
\description{
Draws the evolution of the rate of death of COVID-19 infections based on the
data given. For each gender one graph is shown.
}
\details{
The value at a single point is calculated as the average relative death rate
from the last number of days. The time span which will be used is provided
through the first registered date in the dataframe and number of time steps.
}
\examples{
data <- filter_date(covrki, 20200501, 20201001)

plot_rate_death(data, "A05-A14", 30, 7)
# creates a plot for the subgroup of 5 - 14 year children with a time span of
# 30 days (01.05.2020 - 31.05.2020). The value at a given point is calculated
# as the average rate of death from the last 7 days.

}
