% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_death.R
\name{rate}
\alias{rate}
\title{Calculate rate of death by age and gender}
\usage{
rate(data, age, sex)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble} with RKI data}

\item{age}{age group of which the rate of death is calculated}

\item{sex}{sex of which the rate of death is calculated}
}
\value{
rate of death
}
\description{
Calculate rate of death by age and gender
}
