% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\name{rki_fetch}
\alias{rki_fetch}
\title{Download data from RKI}
\usage{
rki_fetch(from_date, to_date, out_fields = "*")
}
\arguments{
\item{from_date}{Start date (inclusive) of data in YYYYMMDD format}

\item{to_date}{End date (inclusive) of data in YYYYMMDD format}

\item{out_fields}{Columns to include in response}
}
\value{
a \link[tibble:tibble-package]{tibble} containing the requested data
}
\description{
This downloads all data about COVID-19 infections in Germany from the RKI.
This should include data from January 1st, 2020 up to the current date (if
the pandemic is still ongoing). Unfortunately, the API caps the response at
32000 rows which effectively makes this function only useful if you are
interested in data spanning one or two days.
}
\examples{
rki_fetch(20210115, 20210115, c("Altersgruppe", "Geschlecht", "AnzahlFall",
                                "AnzahlTodesfall", "Meldedatum", "NeuerFall",
                                "NeuerTodesfall"))
}
