% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\name{rki_fetch_all}
\alias{rki_fetch_all}
\title{Download all available data from RKI}
\usage{
rki_fetch_all()
}
\value{
a \link[tibble:tibble-package]{tibble} containing all data
}
\description{
This downloads all data about COVID-19 infections in Germany from the RKI.
This should include data from January 1st, 2020 up to the current date (if
the pandemic is still ongoing).
}
