% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\name{rki_json}
\alias{rki_json}
\title{Helper function to obtain the json from RKI API}
\usage{
rki_json(from_date, to_date, out_fields = "*")
}
\arguments{
\item{from_date}{Start date of data in YYYYMMDD format}

\item{to_date}{End date of data in YYYYMMDD format}

\item{out_fields}{Columns to include in response}
}
\value{
a \link[httr:response]{response} to the HTTP GET request to the API
}
\description{
Helper function to obtain the json from RKI API
}
