% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_death.R
\name{table_deaths_by_state}
\alias{table_deaths_by_state}
\title{Summarize deaths by gender and state}
\usage{
table_deaths_by_state(data)
}
\arguments{
\item{data}{\link[tibble:tibble-package]{tibble} with RKI data}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Provides a tibble with the deaths for each state by gender based on
the given data.
}
\examples{
table_deaths_by_state(covrki)
}
