% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YAC19-package.R
\docType{package}
\name{YAC19-package}
\alias{YAC19}
\alias{YAC19-package}
\title{YAC19: Yet Another COVID-19 statistic package}
\description{
Fetch and present statistical data about the number of COVID-19 infections in Germany.
}
\author{
\strong{Maintainer}: Andreas Vollert \email{andreas.vollert@tum.de}

Authors:
\itemize{
  \item Michael Pavlov \email{michael.pavlov@tum.de}
}

}
\keyword{internal}
